___JmolVersion="13.3.1_dev_2013.07.19"
new feature: calculate formalCharge

bug fix: set showTiming not functional
bug fix: set pdbAddHydrogens does not set formal charge on O(1-).
bug fix: PNGJ files not properly caching, and PSE files with DOCACHE option



Former-commit-id: 87ead1e3377a5d271eb63f450aec8b81d05e8830

+++ Jmol/src/org/jmol/adapter/readers/quantum/PsiReader.java
@@ -147,7 +147,7 @@ public class PsiReader extends MOReader {
       if (isInitial) {
         atomNames.addLast(tokens[0]);
         if (tokens[0].length() <= 2)
-          atom.elementNumber = JmolAdapter.getElementNumber(tokens[0]);
+          atom.elementNumber = (short) JmolAdapter.getElementNumber(tokens[0]);
       } else {
         atom.elementNumber = (short) parseIntStr(tokens[0]);
       }
